+++ lib/scss_lint/linter/capitalization_in_selector.rb
@@ -1,0 +1,15 @@
+module SCSSLint
+  class Linter::CapitalizationInSelector < Linter
+    include LinterRegistry
+
+    def visit_rule(node)
+      add_lint(node) if node.rule.first =~ /[A-Z]/
+
+      yield # Continue linting children
+    end
+
+    def description
+      'Selectors should be lowercase'
+    end
+  end
+end