+++ lib/scss_lint/linter/comment.rb
@@ -1,0 +1,13 @@
+module SCSSLint
+  class Linter::Comment < Linter
+    include LinterRegistry
+
+    def visit_comment(node)
+      add_lint(node) unless node.invisible?
+    end
+
+    def description
+      'Use // comments everywhere'
+    end
+  end
+end