+++ lib/scss_lint/linter/empty_rule.rb
@@ -1,0 +1,14 @@
+module SCSSLint
+  class Linter::EmptyRule < Linter
+    include LinterRegistry
+
+    def visit_rule(node)
+      add_lint(node) if node.children.empty?
+      yield # Continue linting children
+    end
+
+    def description
+      'Empty rule'
+    end
+  end
+end