+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Fix bug where using transform-related functions were reported as lints
 * Teach scss-lint that `.5em` is preferred over `0.5em`
 * Teach scss-lint to prefer lowercase characters in selectors
+* Linters names no longer have a `Linter` suffix
 
 0.7.1
 * Upgrade Sass gem dependency from 3.2.8 -> 3.2.9